refactor(native): extract a shared shell-core crate for the desktop and mobile shells
Motivation: the two Tauri shells carried byte-for-byte copies of the
backend-server config model and file format, base-URL normalization,
keyring access-key storage, and the bounded download path — about 400
lines plus their unit tests. They had already drifted once (a trimmed key
authenticated on desktop and failed on mobile), and each side had rules the
other lacked.

Technical changes:
- New Tauri-free crate `native/shell-core` (`openagentd-shell-core`,
  internal, version 0.0.0) with `backend_config`, `access_key`, and
  `download` modules; functions take `&Path` / `&reqwest::Client` and the
  crate asks reqwest only for `json` so each shell keeps its own TLS roots.
- Behaviour unified as the union of both shells: URL normalization strips a
  pasted `/api` (was desktop-only), remove-server compares normalized URLs
  (desktop) and refills the default entry when the list empties (mobile),
  legacy `servers: ["url"]` files are migrated (desktop), download errors
  use the clearer desktop wording with a per-shell timeout.
- `mobile/src/lib.rs` 835 → 408 lines: commands become thin wrappers; only
  runtime state, the reachability gate, iOS sharing, and their tests remain.
- `desktop/src/config.rs` 322 → 99 lines and `commands.rs` 760 → 597: the
  `crate::config` façade re-exports core under the existing names and keeps
  the `AppHandle` adapters, so no desktop call site changed.
- 26 core unit tests (ported from both shells plus new coverage for `/api`
  stripping, normalized removal, legacy migration, oversized base64).
- Both Cargo.lock files updated for the path dependency; no root workspace,
  which would have relocated the lockfiles and broken the `--locked` CI
  contract.
- `make verify-shell-core` (fmt, clippy -D warnings, test) feeds
  `verify-native`; tauri.yml gains a fast Tauri-free shell-core job and adds
  `native/shell-core/**` to both shell jobs' path filters.
- AGENTS.md map/scopes/validation, desktop and mobile guides, and a new
  `native/shell-core/AGENTS.md`.

Impact: −434 net lines; one tested copy of every security-relevant rule.
Verified: shell-core 26 tests, desktop 104 tests + clippy + --locked check,
mobile --locked check + 3 tests.
