+++ mobile/src-tauri/Cargo.toml
@@ -16,6 +16,7 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 tauri-build = { version = "2", features = [] }
 
 [dependencies]
+openagentd-shell-core = { path = "../../native/shell-core" }
 tauri = { version = "2", features = [] }
 time = "=0.3.55"
 tauri-plugin-opener = "2"
