+++ mobile/src-tauri/src/lib.rs
@@ -1,15 +1,21 @@
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Mutex, OnceLock};
 use std::time::Duration;
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{Context, Result};
+use openagentd_shell_core::{
+    load_backend_config_from, normalize_base_url, normalize_server_name,
+    remove_backend_server_at, resolve_download_bytes, safe_download_filename,
+    save_backend_config_to, unique_cache_path, AppBackendConfig, DownloadSource,
+    SavedAppServer,
+};
 use serde::{Deserialize, Serialize};
 use tauri::{AppHandle, Manager};
 
-const MAX_DOWNLOAD_BYTES: usize = 100 * 1024 * 1024;
-const MAX_FILENAME_COLLISION_ATTEMPTS: usize = 1000;
-
-const ACCESS_KEY_SERVICE: &str = "openagentd.backend-access-key";
+/// Total budget for one workspace download. Mobile talks to a remote
+/// server over whatever network the phone has, so keep this shorter than
+/// the desktop's 300 s.
+const DOWNLOAD_TIMEOUT: Duration = Duration::from_secs(60);
 
 /// Process-wide HTTP client shared by the backend health check and file
 /// downloads. Building a `reqwest::Client` per request throws away its
@@ -27,56 +33,22 @@ fn shared_client() -> &'static reqwest::Client {
     })
 }
 
-fn access_key_entry(origin: &str) -> Result<keyring::Entry, String> {
-    let parsed = url::Url::parse(origin).map_err(|_| "invalid backend origin".to_string())?;
-    if !matches!(parsed.scheme(), "http" | "https") {
-        return Err("invalid backend origin".to_string());
-    }
-    let canonical = parsed.origin().ascii_serialization();
-    keyring::Entry::new(ACCESS_KEY_SERVICE, &canonical)
-        .map_err(|_| "credential store unavailable".to_string())
-}
+// Access keys live in the OS credential store, keyed by canonical origin;
+// the logic is shared with the desktop shell in `openagentd-shell-core`.
 
 #[tauri::command]
 fn secure_get_access_key(origin: String) -> Result<Option<String>, String> {
-    match access_key_entry(&origin)?.get_password() {
-        Ok(key) => Ok(Some(key)),
-        Err(keyring::Error::NoEntry) => Ok(None),
-        Err(_) => Err("credential store unavailable".to_string()),
-    }
+    openagentd_shell_core::get_access_key(&origin)
 }
 
 #[tauri::command]
 fn secure_set_access_key(origin: String, key: String) -> Result<(), String> {
-    if key.trim().is_empty() {
-        return Err("access key is required".to_string());
-    }
-    access_key_entry(&origin)?
-        // Store the trimmed value, matching the desktop shell. Validating
-        // `key.trim()` but persisting `key` meant a pasted key with a
-        // trailing newline authenticated on desktop and failed on mobile.
-        .set_password(key.trim())
-        .map_err(|_| "credential store unavailable".to_string())
+    openagentd_shell_core::set_access_key(&origin, &key)
 }
 
 #[tauri::command]
 fn secure_delete_access_key(origin: String) -> Result<(), String> {
-    match access_key_entry(&origin)?.delete_credential() {
-        Ok(()) | Err(keyring::Error::NoEntry) => Ok(()),
-        Err(_) => Err("credential store unavailable".to_string()),
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-struct SavedAppServer {
-    base_url: String,
-    name: Option<String>,
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-struct AppBackendConfig {
-    active_base_url: Option<String>,
-    servers: Vec<SavedAppServer>,
+    openagentd_shell_core::delete_access_key(&origin)
 }
 
 #[derive(Clone, Serialize)]
@@ -98,18 +70,6 @@ fn resolve_active_base_url(
     persisted_base_url: Option<String>,
 ) -> String {
     runtime_base_url.or(persisted_base_url).unwrap_or_default()
-}
-
-impl Default for AppBackendConfig {
-    fn default() -> Self {
-        Self {
-            active_base_url: None,
-            servers: vec![SavedAppServer {
-                base_url: "http://127.0.0.1:4082".to_string(),
-                name: Some("Local CLI server".to_string()),
-            }],
-        }
-    }
 }
 
 #[tauri::command]
@@ -187,7 +147,8 @@ fn app_remove_backend_server(
     base_url: String,
 ) -> Result<AppBackendStatus, String> {
     let normalized = normalize_base_url(&base_url).map_err(|e| format!("{e:#}"))?;
-    remove_backend_server(&app, &normalized).map_err(|e| format!("{e:#}"))?;
+    let path = config_path(&app).map_err(|e| format!("{e:#}"))?;
+    remove_backend_server_at(&path, &normalized).map_err(|e| format!("{e:#}"))?;
     let mut active = state
         .active_base_url
         .lock()
@@ -197,18 +158,6 @@ fn app_remove_backend_server(
     }
     drop(active);
     app_backend_status(app, state)
-}
-
-fn normalize_base_url(base_url: &str) -> Result<String> {
-    let trimmed = base_url.trim().trim_end_matches('/');
-    if trimmed.is_empty() {
-        return Err(anyhow!("base URL is required"));
-    }
-    let parsed = url::Url::parse(trimmed).context("parse base URL")?;
-    match parsed.scheme() {
-        "http" | "https" => Ok(trimmed.to_string()),
-        scheme => Err(anyhow!("unsupported URL scheme: {scheme}")),
-    }
 }
 
 /// Confirm a server answers `GET /api/health/live`.
@@ -228,11 +177,6 @@ async fn ensure_backend_reachable(base_url: &str) -> Result<(), String> {
     Ok(())
 }
 
-fn normalize_server_name(name: Option<String>) -> Option<String> {
-    name.map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
 fn config_path(app: &AppHandle) -> Result<PathBuf> {
     let dir = app
         .path()
@@ -243,8 +187,7 @@ fn config_path(app: &AppHandle) -> Result<PathBuf> {
 }
 
 fn load_backend_config(app: &AppHandle) -> Result<AppBackendConfig> {
-    let path = config_path(app)?;
-    load_backend_config_from(&path)
+    load_backend_config_from(&config_path(app)?)
 }
 
 fn save_backend_config(
@@ -253,105 +196,7 @@ fn save_backend_config(
     name: Option<&str>,
     activate: bool,
 ) -> Result<()> {
-    let path = config_path(app)?;
-    save_backend_config_to(&path, base_url, name, activate)
-}
-
-fn load_backend_config_from(path: &std::path::Path) -> Result<AppBackendConfig> {
-    if !path.exists() {
-        return Ok(AppBackendConfig::default());
-    }
-    let bytes = std::fs::read(path).with_context(|| format!("read {}", path.display()))?;
-    let mut config: AppBackendConfig =
-        serde_json::from_slice(&bytes).with_context(|| format!("parse {}", path.display()))?;
-    if config.servers.is_empty() {
-        config.servers = AppBackendConfig::default().servers;
-    }
-    Ok(config)
-}
-
-fn save_backend_config_to(
-    path: &std::path::Path,
-    base_url: Option<&str>,
-    name: Option<&str>,
-    activate: bool,
-) -> Result<()> {
-    let mut config = load_backend_config_from(path).unwrap_or_default();
-    if activate {
-        config.active_base_url = base_url.map(str::to_string);
-    }
-    if let Some(url) = base_url {
-        if let Some(saved) = config.servers.iter_mut().find(|saved| saved.base_url == url) {
-            if let Some(name) = name {
-                saved.name = Some(name.to_string());
-            }
-        } else {
-            config.servers.push(SavedAppServer {
-                base_url: url.to_string(),
-                name: name.map(str::to_string),
-            });
-        }
-    }
-    let bytes = serde_json::to_vec_pretty(&config).context("serialize backend config")?;
-    std::fs::write(path, bytes).with_context(|| format!("write {}", path.display()))
-}
-
-fn remove_backend_server(app: &AppHandle, base_url: &str) -> Result<()> {
-    let path = config_path(app)?;
-    remove_backend_server_at(&path, base_url)
-}
-
-fn remove_backend_server_at(path: &std::path::Path, base_url: &str) -> Result<()> {
-    let mut config = load_backend_config_from(path).unwrap_or_default();
-    config.servers.retain(|server| server.base_url != base_url);
-    if config.active_base_url.as_deref() == Some(base_url) {
-        config.active_base_url = None;
-    }
-    if config.servers.is_empty() {
-        config.servers = AppBackendConfig::default().servers;
-    }
-    let bytes = serde_json::to_vec_pretty(&config).context("serialize backend config")?;
-    std::fs::write(path, bytes).with_context(|| format!("write {}", path.display()))
-}
-
-fn ensure_max_download_size(size: usize) -> Result<()> {
-    if size > MAX_DOWNLOAD_BYTES {
-        Err(anyhow!("File exceeds 100 MB limit"))
-    } else {
-        Ok(())
-    }
-}
-
-fn ensure_content_length_limit(content_length: Option<u64>) -> Result<()> {
-    if let Some(length) = content_length {
-        let length = usize::try_from(length).map_err(|_| anyhow!("File exceeds 100 MB limit"))?;
-        ensure_max_download_size(length)?;
-    }
-    Ok(())
-}
-
-fn unique_cache_path(dir: &Path, filename: &str) -> Result<PathBuf> {
-    let candidate = dir.join(filename);
-    if !candidate.exists() {
-        return Ok(candidate);
-    }
-
-    let path = Path::new(filename);
-    let stem = path.file_stem().and_then(|value| value.to_str()).unwrap_or(filename);
-    let ext = path.extension().and_then(|value| value.to_str());
-
-    for index in 1..=MAX_FILENAME_COLLISION_ATTEMPTS {
-        let candidate_name = match ext {
-            Some(ext) if !ext.is_empty() => format!("{stem} ({index}).{ext}"),
-            _ => format!("{stem} ({index})"),
-        };
-        let candidate = dir.join(candidate_name);
-        if !candidate.exists() {
-            return Ok(candidate);
-        }
-    }
-
-    Err(anyhow!("Could not find a unique filename after 1000 attempts"))
+    save_backend_config_to(&config_path(app)?, base_url, name, activate)
 }
 
 #[derive(Deserialize)]
@@ -368,83 +213,16 @@ async fn save_workspace_file(
     app: AppHandle,
     request: SaveWorkspaceFileRequest,
 ) -> Result<bool, String> {
-    let filename = Path::new(&request.filename)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .filter(|name| !name.is_empty())
-        .unwrap_or("download")
-        .to_string();
-
-    // Resolve the file bytes: base64 payload, data: URI, or remote URL fetch.
-    let bytes: Vec<u8> = match (request.base64, request.url) {
-        (Some(b64), _) => {
-            use base64::Engine;
-            let bytes = base64::prelude::BASE64_STANDARD
-                .decode(&b64)
-                .map_err(|e| format!("Decode base64: {e}"))?;
-            ensure_max_download_size(bytes.len()).map_err(|e| format!("Decode base64: {e}"))?;
-            bytes
-        }
-        (None, Some(url)) if url.starts_with("data:") => {
-            // data:<mime>;base64,<payload>
-            use base64::Engine;
-            let payload = url
-                .split_once(',')
-                .map(|(_, payload)| payload)
-                .ok_or("Invalid data URI: missing comma")?;
-            let bytes = base64::prelude::BASE64_STANDARD
-                .decode(payload)
-                .map_err(|e| format!("Decode data URI: {e}"))?;
-            ensure_max_download_size(bytes.len()).map_err(|e| format!("Decode data URI: {e}"))?;
-            bytes
-        }
-        (None, Some(url)) => {
-            let mut request_builder = shared_client()
-                .get(&url)
-                .timeout(Duration::from_secs(60));
-            if let Ok(parsed) = url::Url::parse(&url) {
-                let has_token_qs = parsed.query_pairs().any(|(k, _)| k == "_token");
-                if !has_token_qs {
-                    let origin = parsed.origin().ascii_serialization();
-                    if let Ok(Some(key)) = secure_get_access_key(origin) {
-                        request_builder = request_builder.bearer_auth(key);
-                    }
-                }
-            }
-            let mut response = request_builder
-                .send()
-                .await
-                .map_err(|e| format!("Fetch file: {e}"))?
-                .error_for_status()
-                .map_err(|e| format!("Fetch file: {e}"))?;
-            ensure_content_length_limit(response.content_length())
-                .map_err(|e| format!("Fetch file: {e}"))?;
-
-            // Pre-reserve from the (already cap-checked) Content-Length so
-            // a large body doesn't pay repeated doubling reallocations; the
-            // `min` keeps a lying header from over-allocating.
-            let mut bytes = Vec::with_capacity(
-                response
-                    .content_length()
-                    .map_or(0, |length| length.min(MAX_DOWNLOAD_BYTES as u64) as usize),
-            );
-            while let Some(chunk) = response
-                .chunk()
-                .await
-                .map_err(|e| format!("Read fetched file: {e}"))?
-            {
-                let new_len = bytes
-                    .len()
-                    .checked_add(chunk.len())
-                    .ok_or_else(|| "Read fetched file: File exceeds 100 MB limit".to_string())?;
-                ensure_max_download_size(new_len)
-                    .map_err(|e| format!("Read fetched file: {e}"))?;
-                bytes.extend_from_slice(&chunk);
-            }
-            bytes
-        }
-        (None, None) => return Err("save_workspace_file: must supply either base64 or url".to_string()),
-    };
+    let filename = safe_download_filename(&request.filename);
+    let bytes = resolve_download_bytes(
+        shared_client(),
+        DownloadSource {
+            base64: request.base64.as_deref(),
+            url: request.url.as_deref(),
+        },
+        DOWNLOAD_TIMEOUT,
+    )
+    .await?;
 
     let cache_dir = app
         .path()
@@ -587,46 +365,11 @@ pub fn run() {
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        ensure_backend_reachable, ensure_max_download_size, load_backend_config_from,
-        normalize_base_url, normalize_server_name, remove_backend_server_at,
-        resolve_active_base_url, save_backend_config_to, unique_cache_path, AppBackendConfig,
-        MAX_DOWNLOAD_BYTES,
-    };
-    use std::path::PathBuf;
-    use tempfile::tempdir;
+    use super::{ensure_backend_reachable, resolve_active_base_url};
 
-    fn config_file() -> (tempfile::TempDir, PathBuf) {
-        let dir = tempdir().expect("tempdir");
-        let path = dir.path().join("app-backend.json");
-        (dir, path)
-    }
-
-    #[test]
-    fn normalize_base_url_accepts_http_and_https() {
-        assert_eq!(normalize_base_url("http://example.com").unwrap(), "http://example.com");
-        assert_eq!(normalize_base_url("https://example.com/path").unwrap(), "https://example.com/path");
-    }
-
-    #[test]
-    fn normalize_base_url_trims_whitespace_and_trailing_slashes() {
-        assert_eq!(normalize_base_url("  https://example.com/path///  ").unwrap(), "https://example.com/path");
-    }
-
-    #[test]
-    fn normalize_base_url_rejects_empty_scheme_and_invalid_urls() {
-        assert!(normalize_base_url("   ").is_err());
-        assert!(normalize_base_url("file:///tmp/test").is_err());
-        assert!(normalize_base_url("ftp://example.com").is_err());
-        assert!(normalize_base_url("not a url").is_err());
-    }
-
-    #[test]
-    fn normalize_server_name_trims_and_drops_empty_values() {
-        assert_eq!(normalize_server_name(Some("  My Server  ".to_string())), Some("My Server".to_string()));
-        assert_eq!(normalize_server_name(Some("   ".to_string())), None);
-        assert_eq!(normalize_server_name(None), None);
-    }
+    // Everything about config files, URL normalisation, access keys, and
+    // download limits is tested once in `openagentd-shell-core`; only the
+    // mobile-specific runtime state and reachability gate live here.
 
     #[test]
     fn runtime_backend_overrides_persisted_startup_backend() {
@@ -645,176 +388,6 @@ mod tests {
             resolve_active_base_url(None, Some("https://persisted.example".to_string())),
             "https://persisted.example",
         );
-    }
-
-    #[test]
-    fn load_backend_config_returns_default_when_missing() {
-        let (_dir, path) = config_file();
-
-        let config = load_backend_config_from(&path).unwrap();
-
-        assert_eq!(config.active_base_url, None);
-        assert_eq!(config.servers.len(), 1);
-        assert_eq!(config.servers[0].base_url, "http://127.0.0.1:4082");
-        assert_eq!(config.servers[0].name.as_deref(), Some("Local CLI server"));
-    }
-
-    #[test]
-    fn save_backend_config_adds_new_server_without_activating() {
-        let (_dir, path) = config_file();
-
-        save_backend_config_to(&path, Some("https://example.com"), Some("Example"), false)
-            .unwrap();
-        let config = load_backend_config_from(&path).unwrap();
-
-        assert_eq!(config.active_base_url, None);
-        assert!(config.servers.iter().any(|server| {
-            server.base_url == "https://example.com" && server.name.as_deref() == Some("Example")
-        }));
-    }
-
-    #[test]
-    fn save_backend_config_can_activate_server() {
-        let (_dir, path) = config_file();
-
-        save_backend_config_to(&path, Some("https://example.com"), Some("Example"), true)
-            .unwrap();
-        let config = load_backend_config_from(&path).unwrap();
-
-        assert_eq!(config.active_base_url.as_deref(), Some("https://example.com"));
-    }
-
-    #[test]
-    fn save_backend_config_updates_existing_server_name() {
-        let (_dir, path) = config_file();
-
-        save_backend_config_to(&path, Some("https://example.com"), Some("Before"), false)
-            .unwrap();
-        save_backend_config_to(&path, Some("https://example.com"), Some("After"), false)
-            .unwrap();
-        let config = load_backend_config_from(&path).unwrap();
-        let matching: Vec<_> = config
-            .servers
-            .iter()
-            .filter(|server| server.base_url == "https://example.com")
-            .collect();
-
-        assert_eq!(matching.len(), 1);
-        assert_eq!(matching[0].name.as_deref(), Some("After"));
-    }
-
-    #[test]
-    fn remove_backend_server_clears_active_base_url_when_matching() {
-        let (_dir, path) = config_file();
-
-        save_backend_config_to(&path, Some("https://example.com"), Some("Example"), true)
-            .unwrap();
-        remove_backend_server_at(&path, "https://example.com").unwrap();
-        let config = load_backend_config_from(&path).unwrap();
-
-        assert_eq!(config.active_base_url, None);
-        assert!(!config.servers.iter().any(|server| server.base_url == "https://example.com"));
-    }
-
-    #[test]
-    fn empty_servers_list_falls_back_to_default_on_load_and_after_remove() {
-        let (_dir, path) = config_file();
-
-        std::fs::write(
-            &path,
-            serde_json::to_vec_pretty(&AppBackendConfig {
-                active_base_url: Some("https://example.com".to_string()),
-                servers: vec![],
-            })
-            .unwrap(),
-        )
-        .unwrap();
-
-        let loaded = load_backend_config_from(&path).unwrap();
-        assert_eq!(loaded.servers.len(), 1);
-        assert_eq!(loaded.servers[0].base_url, "http://127.0.0.1:4082");
-
-        save_backend_config_to(&path, Some("https://example.com"), Some("Example"), true)
-            .unwrap();
-        remove_backend_server_at(&path, "https://example.com").unwrap();
-        let after_remove = load_backend_config_from(&path).unwrap();
-
-        assert_eq!(after_remove.active_base_url, None);
-        assert_eq!(after_remove.servers.len(), 1);
-        assert_eq!(after_remove.servers[0].base_url, "http://127.0.0.1:4082");
-    }
-
-    #[test]
-    fn load_backend_config_rejects_corrupt_json() {
-        let (_dir, path) = config_file();
-        std::fs::write(&path, b"{not json").unwrap();
-
-        assert!(load_backend_config_from(&path).is_err());
-    }
-
-    #[test]
-    fn unique_cache_path_returns_original_when_no_collision() {
-        let dir = tempdir().unwrap();
-
-        let path = unique_cache_path(dir.path(), "report.txt").unwrap();
-
-        assert_eq!(path, dir.path().join("report.txt"));
-    }
-
-    #[test]
-    fn unique_cache_path_adds_suffix_for_single_collision() {
-        let dir = tempdir().unwrap();
-        std::fs::write(dir.path().join("report.txt"), b"existing").unwrap();
-
-        let path = unique_cache_path(dir.path(), "report.txt").unwrap();
-
-        assert_eq!(path, dir.path().join("report (1).txt"));
-    }
-
-    #[test]
-    fn unique_cache_path_skips_to_next_available_suffix() {
-        let dir = tempdir().unwrap();
-        std::fs::write(dir.path().join("report.txt"), b"existing").unwrap();
-        std::fs::write(dir.path().join("report (1).txt"), b"existing").unwrap();
-        std::fs::write(dir.path().join("report (2).txt"), b"existing").unwrap();
-
-        let path = unique_cache_path(dir.path(), "report.txt").unwrap();
-
-        assert_eq!(path, dir.path().join("report (3).txt"));
-    }
-
-    #[test]
-    fn unique_cache_path_preserves_extension() {
-        let dir = tempdir().unwrap();
-        std::fs::write(dir.path().join("archive.tar.gz"), b"existing").unwrap();
-
-        let path = unique_cache_path(dir.path(), "archive.tar.gz").unwrap();
-
-        assert_eq!(path, dir.path().join("archive.tar (1).gz"));
-    }
-
-    #[test]
-    fn unique_cache_path_handles_names_without_extension() {
-        let dir = tempdir().unwrap();
-        std::fs::write(dir.path().join("download"), b"existing").unwrap();
-
-        let path = unique_cache_path(dir.path(), "download").unwrap();
-
-        assert_eq!(path, dir.path().join("download (1)"));
-    }
-
-    #[test]
-    fn access_key_origin_must_be_a_canonical_http_origin() {
-        assert!(super::access_key_entry("https://example.com").is_ok());
-        assert!(super::access_key_entry("https://example.com/api").is_ok());
-        assert!(super::access_key_entry("ftp://example.com").is_err());
-        assert!(super::access_key_entry("https://example.com/").is_ok());
-    }
-
-    #[test]
-    fn ensure_max_download_size_rejects_oversized_payload() {
-        assert!(ensure_max_download_size(MAX_DOWNLOAD_BYTES).is_ok());
-        assert!(ensure_max_download_size(MAX_DOWNLOAD_BYTES + 1).is_err());
     }
 
     // ── ensure_backend_reachable ────────────────────────────────────────
