+++ native/shell-core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "openagentd-shell-core"
+version = "0.0.0"
+description = "Tauri-free logic shared by the OpenAgentd desktop and mobile shells"
+authors = ["openagentd contributors"]
+license = "Apache-2.0"
+repository = "https://github.com/lthoangg/openagentd"
+edition = "2021"
+rust-version = "1.77.2"
+publish = false
+
+[dependencies]
+anyhow = "1"
+base64 = "0.23"
+keyring = { version = "3.6.3", default-features = false }
+# TLS backend and platform roots are chosen by each shell's own reqwest
+# dependency; Cargo unifies features across the graph, so this crate only
+# asks for what it calls.
+reqwest = { version = "0.12", features = ["json"], default-features = false }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+url = "2"
+
+[dev-dependencies]
+tempfile = "3"
+tokio = { version = "1", features = ["macros", "rt"] }
